feat: [TestRun] moved Scenario types into scenario/Types.tsx
