+++ frontend/src/components/scenario/Types.tsx
@@ -1,4 +1,12 @@
-import {Scenario} from "../../Types.tsx";
+export type Scenario = {
+    id       : string
+    authorID : string
+    label    : string
+}
+
+export type NewScenario = {
+    label    : string
+}
 
 export type ScenarioDialogOptions = {
     scenario: Scenario
